feat: add FolderCard and PasswordCard components, and integrate them into PasswordsList
