+++ frontend/components.d.ts
@@ -17,6 +17,7 @@ declare module 'vue' {
     Dialog: typeof import('primevue/dialog')['default']
     Divider: typeof import('primevue/divider')['default']
     Drawer: typeof import('primevue/drawer')['default']
+    FolderCard: typeof import('./src/components/passwords/FolderCard.vue')['default']
     Form: typeof import('@primevue/forms/form')['default']
     InputNumber: typeof import('primevue/inputnumber')['default']
     InputText: typeof import('primevue/inputtext')['default']
@@ -25,6 +26,8 @@ declare module 'vue' {
     Message: typeof import('primevue/message')['default']
     PanelMenu: typeof import('primevue/panelmenu')['default']
     Password: typeof import('primevue/password')['default']
+    PasswordCard: typeof import('./src/components/passwords/PasswordCard.vue')['default']
+    PasswordsList: typeof import('./src/components/passwords/PasswordsList.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
     SelectButton: typeof import('primevue/selectbutton')['default']
