+++ frontend/src/components/passwords/PasswordsList.vue
@@ -1,0 +1,65 @@
+<template>
+  <div>
+    <div v-if="loading" class="text-center py-8">
+      <p>Loading passwords...</p>
+    </div>
+    
+    <div v-else-if="error" class="bg-red-100 dark:bg-red-900 border border-red-400 dark:border-red-600 text-red-700 dark:text-red-200 px-4 py-3 rounded mb-4">
+      {{ error }}
+    </div>
+    
+    <div v-else-if="passwords.length === 0" class="text-center py-8 text-gray-500 dark:text-gray-400">
+      <p>No passwords found. Create your first password!</p>
+    </div>
+    
+    <div v-else class="space-y-2">
+      <FolderCard
+        v-for="folder in folders"
+        :key="folder.name"
+        :folder="folder"
+        :initialOpen="selectedFolder === folder.name"
+      />
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import { computed } from 'vue';
+import type { GetPasswordListResponse } from '@/client/types.gen';
+import FolderCard from './FolderCard.vue';
+
+const props = defineProps<{
+  passwords: GetPasswordListResponse[];
+  loading: boolean;
+  error: string | null;
+  selectedFolder: string | null;
+  folderFilter?: string | null;
+}>();
+
+// Group passwords by folder
+const folders = computed(() => {
+  const folderMap = new Map<string, GetPasswordListResponse[]>();
+  
+  props.passwords.forEach(password => {
+    const folderName = password.folder || 'default';
+    if (!folderMap.has(folderName)) {
+      folderMap.set(folderName, []);
+    }
+    folderMap.get(folderName)!.push(password);
+  });
+  
+  // Filter by folder query param if present
+  if (props.folderFilter) {
+    const filtered = folderMap.get(props.folderFilter);
+    if (filtered) {
+      return [{ name: props.folderFilter, count: filtered.length, passwords: filtered }];
+    }
+  }
+  
+  return Array.from(folderMap.entries()).map(([name, items]) => ({
+    name,
+    count: items.length,
+    passwords: items
+  }));
+});
+</script>
